requirements.txt
aioredis==1.3.1
amqp==5.0.9
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
billiard==3.6.4.0
boto3==1.20.42
botocore==1.23.42
Brotli==1.0.9
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.4.1
daphne==3.0.2
Deprecated==1.2.13
Django==3.2
django-admin-interface==0.18.5
django-admin-rangefilter==0.8.3
django-admin-views==0.8.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-colorfield==0.6.3
django-cors-headers==3.11.0
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-geojson==3.2.0
django-json-widget==1.1.1
django-leaflet==0.28.2
django-lifecycle==0.9.3
django-map-widgets==0.3.2
django-nine==0.2.5
django-rest-elasticsearch==0.4.2
django-smtp-ssl==1.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-gis==0.18
drf-yasg==1.20.0
elasticsearch==7.12.0
elasticsearch-dsl==7.3.0
fonttools==4.29.0
future==0.18.2
googlemaps==4.5.3
h11==0.12.0
hiredis==2.0.0
html5lib==1.1
httpcore==0.14.5
httpx==0.21.3
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.10.1
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
kombu==5.2.3
Markdown==3.3.6
MarkupSafe==2.0.1
msgpack==1.0.3
packaging==21.3
Pillow==8.0.0
prompt-toolkit==3.0.24
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydyf==0.1.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyphen==0.12.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.1
requests==2.27.1
rfc3986==1.5.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
sentry-sdk==1.5.4
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
tinycss2==1.1.1
twilio==6.38.0
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
urlman==2.0.1
vine==5.0.0
wcwidth==0.2.5
weasyprint==54.0
webencodings==0.5.1
wrapt==1.13.3
xlrd==1.2.0
xlwt==1.3.0
zipp==3.7.0
zope.interface==5.4.0
zopfli==0.1.9
dj-static
pydantic
static
static3
uWSGI
